Add initial fake-database
Buggy maddening monstrous amalgamating calmaity of half-server
half-client side madness that I don't want to deal with anymore at this
hour. Probably going to just pull the JSON file from the phonegap server
at app startup, but I'd rather not have more contact after that.

Why? Because that means writing a lot of server-side code, and I don't
want to do that. :\

Why not write server-side code? Because I have yet to learn it that much
or all that well.